% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{baseballmexR-package}
\alias{baseballmexR}
\alias{baseballmexR-package}
\title{baseballmexR: Query MLB stats api for Mexican Baseball Pro Leagues}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/axelmora/baseballmexR}
  \item Report bugs at \url{https://github.com/axelmora/baseballmexR/issues}
}

}
\author{
\strong{Maintainer}: Axel Mora \email{axel.asmo@gmail.com}

}
\keyword{internal}
