% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule}
\alias{schedule}
\title{\strong{Find game_pk values for Mexican professional baseball games (LMB and LMP)}}
\usage{
schedule(season = 2024, league = c("lmb", "lmp"), game_type = "R")
}
\arguments{
\item{season}{The season for which you want to find game_pk values for MLB games}

\item{league}{lmb or lmp}

\item{game_type}{character vector of game types to include. Options are "R" (regular sesason),
"F" (LMB first-round playoff series), "D" (LMB zone series, LMP first round), "L" (LMB zone
championship series, LMP semi-final), "W" (LMB Serie del Rey, LMP Serie Final)
Default is "R".}
}
\value{
Returns a tibble which includes \code{game_pk} values and additional
information for games scheduled or played with the following columns:\tabular{ll}{
   col_name \tab types \cr
   date \tab character \cr
   total_items \tab integer \cr
   total_events \tab integer \cr
   total_games \tab integer \cr
   total_games_in_progress \tab integer \cr
   game_pk \tab integer \cr
   link \tab character \cr
   game_type \tab character \cr
   season \tab character \cr
   game_date \tab character \cr
   official_date \tab character \cr
   game_number \tab integer \cr
   public_facing \tab logical \cr
   double_header \tab character \cr
   gameday_type \tab character \cr
   tiebreaker \tab character \cr
   calendar_event_id \tab character \cr
   season_display \tab character \cr
   day_night \tab character \cr
   scheduled_innings \tab integer \cr
   reverse_home_away_status \tab logical \cr
   inning_break_length \tab integer \cr
   games_in_series \tab integer \cr
   series_game_number \tab integer \cr
   series_description \tab character \cr
   record_source \tab character \cr
   if_necessary \tab character \cr
   if_necessary_description \tab character \cr
   status_abstract_game_state \tab character \cr
   status_coded_game_state \tab character \cr
   status_detailed_state \tab character \cr
   status_status_code \tab character \cr
   status_start_time_tbd \tab logical \cr
   status_reason \tab character \cr
   status_abstract_game_code \tab character \cr
   teams_away_split_squad \tab logical \cr
   teams_away_series_number \tab integer \cr
   teams_away_league_record_wins \tab integer \cr
   teams_away_league_record_losses \tab integer \cr
   teams_away_league_record_pct \tab character \cr
   teams_away_team_id \tab integer \cr
   teams_away_team_name \tab character \cr
   teams_away_team_link \tab character \cr
   teams_home_split_squad \tab logical \cr
   teams_home_series_number \tab integer \cr
   teams_home_league_record_wins \tab integer \cr
   teams_home_league_record_losses \tab integer \cr
   teams_home_league_record_pct \tab character \cr
   teams_home_team_id \tab integer \cr
   teams_home_team_name \tab character \cr
   teams_home_team_link \tab character \cr
   venue_id \tab integer \cr
   venue_name \tab character \cr
   venue_link \tab character \cr
   content_link \tab character \cr
   is_tie \tab logical \cr
   description \tab character \cr
   teams_away_score \tab integer \cr
   teams_away_is_winner \tab logical \cr
   teams_home_score \tab integer \cr
   teams_home_is_winner \tab logical \cr
   reschedule_date \tab character \cr
   reschedule_game_date \tab character \cr
   rescheduled_from \tab character \cr
   rescheduled_from_date \tab character \cr
   resume_date \tab character \cr
   resume_game_date \tab character \cr
   resumed_from \tab character \cr
   resumed_from_date \tab character \cr
   events \tab list \cr
}
}
\description{
\strong{Find game_pk values for Mexican professional baseball games (LMB and LMP)}
}
\section{Level IDs}{

}

\examples{
\donttest{
  try(mlb_schedule(season = "2019"))
}
}
